move some admin function to ddl
