+++ store/gcworker/gc_worker.go
@@ -34,6 +34,7 @@ import (
 	"github.com/pingcap/kvproto/pkg/errorpb"
 	"github.com/pingcap/kvproto/pkg/kvrpcpb"
 	"github.com/pingcap/kvproto/pkg/metapb"
+	"github.com/pingcap/tidb/ddl"
 	"github.com/pingcap/tidb/ddl/label"
 	"github.com/pingcap/tidb/ddl/placement"
 	"github.com/pingcap/tidb/ddl/util"
@@ -47,7 +48,6 @@ import (
 	"github.com/pingcap/tidb/session"
 	"github.com/pingcap/tidb/sessionctx/variable"
 	"github.com/pingcap/tidb/tablecodec"
-	"github.com/pingcap/tidb/util/admin"
 	"github.com/pingcap/tidb/util/codec"
 	"github.com/pingcap/tidb/util/logutil"
 	tikverr "github.com/tikv/client-go/v2/error"
@@ -1891,7 +1891,7 @@ func (w *GCWorker) doGCPlacementRules(safePoint uint64, dr util.DelRangeTask, gc
 			return
 		}
 		if historyJob == nil {
-			return admin.ErrDDLJobNotFound.GenWithStackByArgs(dr.JobID)
+			return ddl.ErrDDLJobNotFound.GenWithStackByArgs(dr.JobID)
 		}
 	}
 
@@ -1966,7 +1966,7 @@ func (w *GCWorker) doGCLabelRules(dr util.DelRangeTask) (err error) {
 			return
 		}
 		if historyJob == nil {
-			return admin.ErrDDLJobNotFound.GenWithStackByArgs(dr.JobID)
+			return ddl.ErrDDLJobNotFound.GenWithStackByArgs(dr.JobID)
 		}
 	}
