+++ util/admin/admin.go
@@ -16,16 +16,13 @@ package admin
 
 import (
 	"context"
-	"encoding/json"
 	"math"
-	"sort"
 	"strings"
 
 	"github.com/pingcap/errors"
 	"github.com/pingcap/tidb/errno"
 	"github.com/pingcap/tidb/expression"
 	"github.com/pingcap/tidb/kv"
-	"github.com/pingcap/tidb/meta"
 	"github.com/pingcap/tidb/parser/model"
 	"github.com/pingcap/tidb/parser/mysql"
 	"github.com/pingcap/tidb/sessionctx"
@@ -40,218 +37,6 @@ import (
 	"github.com/pingcap/tidb/util/sqlexec"
 	"go.uber.org/zap"
 )
-
-// DDLInfo is for DDL information.
-type DDLInfo struct {
-	SchemaVer   int64
-	ReorgHandle kv.Key       // It's only used for DDL information.
-	Jobs        []*model.Job // It's the currently running jobs.
-}
-
-// GetDDLInfo returns DDL information.
-func GetDDLInfo(txn kv.Transaction) (*DDLInfo, error) {
-	var err error
-	info := &DDLInfo{}
-	t := meta.NewMeta(txn)
-
-	info.Jobs = make([]*model.Job, 0, 2)
-	job, err := t.GetDDLJobByIdx(0)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	if job != nil {
-		info.Jobs = append(info.Jobs, job)
-	}
-	addIdxJob, err := t.GetDDLJobByIdx(0, meta.AddIndexJobListKey)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	if addIdxJob != nil {
-		info.Jobs = append(info.Jobs, addIdxJob)
-	}
-
-	info.SchemaVer, err = t.GetSchemaVersion()
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	if addIdxJob == nil {
-		return info, nil
-	}
-
-	_, info.ReorgHandle, _, _, err = t.GetDDLReorgHandle(addIdxJob)
-	if err != nil {
-		if meta.ErrDDLReorgElementNotExist.Equal(err) {
-			return info, nil
-		}
-		return nil, errors.Trace(err)
-	}
-
-	return info, nil
-}
-
-// CancelJobs cancels the DDL jobs.
-func CancelJobs(txn kv.Transaction, ids []int64) ([]error, error) {
-	if len(ids) == 0 {
-		return nil, nil
-	}
-
-	errs := make([]error, len(ids))
-	t := meta.NewMeta(txn)
-	generalJobs, err := getDDLJobsInQueue(t, meta.DefaultJobListKey)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	addIdxJobs, err := getDDLJobsInQueue(t, meta.AddIndexJobListKey)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	jobs := append(generalJobs, addIdxJobs...)
-	jobsMap := make(map[int64]int)
-	for i, id := range ids {
-		jobsMap[id] = i
-	}
-	for j, job := range jobs {
-		i, ok := jobsMap[job.ID]
-		if !ok {
-			logutil.BgLogger().Debug("the job that needs to be canceled isn't equal to current job",
-				zap.Int64("need to canceled job ID", job.ID),
-				zap.Int64("current job ID", job.ID))
-			continue
-		}
-		delete(jobsMap, job.ID)
-		// These states can't be cancelled.
-		if job.IsDone() || job.IsSynced() {
-			errs[i] = ErrCancelFinishedDDLJob.GenWithStackByArgs(job.ID)
-			continue
-		}
-		// If the state is rolling back, it means the work is cleaning the data after cancelling the job.
-		if job.IsCancelled() || job.IsRollingback() || job.IsRollbackDone() {
-			continue
-		}
-		if !job.IsRollbackable() {
-			errs[i] = ErrCannotCancelDDLJob.GenWithStackByArgs(job.ID)
-			continue
-		}
-
-		job.State = model.JobStateCancelling
-		// Make sure RawArgs isn't overwritten.
-		err := json.Unmarshal(job.RawArgs, &job.Args)
-		if err != nil {
-			errs[i] = errors.Trace(err)
-			continue
-		}
-		if j >= len(generalJobs) {
-			offset := int64(j - len(generalJobs))
-			err = t.UpdateDDLJob(offset, job, true, meta.AddIndexJobListKey)
-		} else {
-			err = t.UpdateDDLJob(int64(j), job, true)
-		}
-		if err != nil {
-			errs[i] = errors.Trace(err)
-		}
-	}
-	for id, i := range jobsMap {
-		errs[i] = ErrDDLJobNotFound.GenWithStackByArgs(id)
-	}
-	return errs, nil
-}
-
-func getDDLJobsInQueue(t *meta.Meta, jobListKey meta.JobListKeyType) ([]*model.Job, error) {
-	cnt, err := t.DDLJobQueueLen(jobListKey)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	jobs := make([]*model.Job, cnt)
-	for i := range jobs {
-		jobs[i], err = t.GetDDLJobByIdx(int64(i), jobListKey)
-		if err != nil {
-			return nil, errors.Trace(err)
-		}
-	}
-	return jobs, nil
-}
-
-// GetDDLJobs get all DDL jobs and sorts jobs by job.ID.
-func GetDDLJobs(txn kv.Transaction) ([]*model.Job, error) {
-	t := meta.NewMeta(txn)
-	generalJobs, err := getDDLJobsInQueue(t, meta.DefaultJobListKey)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	addIdxJobs, err := getDDLJobsInQueue(t, meta.AddIndexJobListKey)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	jobs := append(generalJobs, addIdxJobs...)
-	sort.Sort(jobArray(jobs))
-	return jobs, nil
-}
-
-type jobArray []*model.Job
-
-func (v jobArray) Len() int {
-	return len(v)
-}
-
-func (v jobArray) Less(i, j int) bool {
-	return v[i].ID < v[j].ID
-}
-
-func (v jobArray) Swap(i, j int) {
-	v[i], v[j] = v[j], v[i]
-}
-
-// MaxHistoryJobs is exported for testing.
-const MaxHistoryJobs = 10
-
-// DefNumHistoryJobs is default value of the default number of history job
-const DefNumHistoryJobs = 10
-
-// GetHistoryDDLJobs returns the DDL history jobs and an error.
-// The maximum count of history jobs is num.
-func GetHistoryDDLJobs(txn kv.Transaction, maxNumJobs int) ([]*model.Job, error) {
-	t := meta.NewMeta(txn)
-	jobs, err := t.GetLastNHistoryDDLJobs(maxNumJobs)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	return jobs, nil
-}
-
-// IterHistoryDDLJobs iterates history DDL jobs until the `finishFn` return true or error.
-func IterHistoryDDLJobs(txn kv.Transaction, finishFn func([]*model.Job) (bool, error)) error {
-	txnMeta := meta.NewMeta(txn)
-	iter, err := txnMeta.GetLastHistoryDDLJobsIterator()
-	if err != nil {
-		return err
-	}
-	cacheJobs := make([]*model.Job, 0, DefNumHistoryJobs)
-	for {
-		cacheJobs, err = iter.GetLastJobs(DefNumHistoryJobs, cacheJobs)
-		if err != nil || len(cacheJobs) == 0 {
-			return err
-		}
-		finish, err := finishFn(cacheJobs)
-		if err != nil || finish {
-			return err
-		}
-	}
-}
-
-// IterAllDDLJobs will iterates running DDL jobs first, return directly if `finishFn` return true or error,
-// then iterates history DDL jobs until the `finishFn` return true or error.
-func IterAllDDLJobs(txn kv.Transaction, finishFn func([]*model.Job) (bool, error)) error {
-	jobs, err := GetDDLJobs(txn)
-	if err != nil {
-		return err
-	}
-
-	finish, err := finishFn(jobs)
-	if err != nil || finish {
-		return err
-	}
-	return IterHistoryDDLJobs(txn, finishFn)
-}
 
 // RecordData is the record data composed of a handle and values.
 type RecordData struct {
@@ -472,12 +257,6 @@ func iterRecords(sessCtx sessionctx.Context, retriever kv.Retriever, t table.Tab
 }
 
 var (
-	// ErrDDLJobNotFound indicates the job id was not found.
-	ErrDDLJobNotFound = dbterror.ClassAdmin.NewStd(errno.ErrDDLJobNotFound)
-	// ErrCancelFinishedDDLJob returns when cancel a finished ddl job.
-	ErrCancelFinishedDDLJob = dbterror.ClassAdmin.NewStd(errno.ErrCancelFinishedDDLJob)
-	// ErrCannotCancelDDLJob returns when cancel a almost finished ddl job, because cancel in now may cause data inconsistency.
-	ErrCannotCancelDDLJob = dbterror.ClassAdmin.NewStd(errno.ErrCannotCancelDDLJob)
 	// ErrAdminCheckTable returns when the table records is inconsistent with the index values.
 	ErrAdminCheckTable = dbterror.ClassAdmin.NewStd(errno.ErrAdminCheckTable)
 )
